+++ app/Http/Controllers/TopicController.php
@@ -9,6 +9,7 @@ use App\Services\UserService;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
 
 class TopicController extends Controller
 {
@@ -23,28 +24,36 @@ class TopicController extends Controller
     public function index(){
         $topics = $this->userService->gettopicsForUser(auth()->user())->paginate(10);
         $topics->load('subject', 'questions');
-        $header = ['ID', 'Subject', 'Name', 'Question Count', 'Date Created'];
+        $header = ['Subject', 'Name', 'Question Count'];
         $rows = $topics->map(function ($topic) {
             return [
                 'id' => $topic->id,
-                'subject' => $topic->subject->name,
+                'subject' => $topic->subject->code,
                 'name' => $topic->name,
-                'question count' => $topic->questions->count(),
-                'Date Created' => Carbon::parse($topic->created_at)->format('m/d/Y')
+                'question count' => $topic->questions->count()
             ];
         });
 
         $data = [
             'headers' => $header,
             'rows' => $rows,
-            'topics' => $topics
+            'models' => $topics,
+            'url' => 'topics'
         ];
+
+        if (request()->hasHeader('HX-Request') && !request()->hasHeader('HX-History-Restore-Request')) {
+            // Return only the partial view for HTMX
+            return view('components/core/index-table', $data);
+        }
 
         return view('topics/index', $data);
     }
 
     public function show(Topic $topic){
-        $header = ['ID', 'Name', 'Type', 'Date Created'];
+        $questions_are_in_exams = $topic->questions->contains(function ($question) {
+            return $question->exams()->exists();
+        });    
+        $header = ['Name', 'Type', 'Date Created'];
         $questions = Question::with(['topic'])
             ->where('topic_id', $topic->id)
             ->Paginate(5);
@@ -59,8 +68,15 @@ class TopicController extends Controller
             'headers' => $header,
             'rows' => $rows,
             'topic'=>$topic,
-            'questions' => $questions
+            'models' => $questions,
+            'url' => 'questions',
+            'questions_are_in_exams' => $questions_are_in_exams
         ];
+
+        if (request()->hasHeader('HX-Request') && !request()->hasHeader('HX-History-Restore-Request')) {
+            // Return only the partial view for HTMX
+            return view('components/core/index-table', $data);
+        }
 
         return view('topics/show', $data);
     }
@@ -71,8 +87,9 @@ class TopicController extends Controller
     }
 
     public function store(){
+        $subject_id = request()->input('subject');
         $validator = Validator::make(request()->post(), [
-            'name'    => ['required'],
+            'name'    => ['required', 'unique:topics,name,NULL,id,subject_id,' . $subject_id],
             'subject' => ['required', 'integer', 'exists:subjects,id'],
         ]);
 
@@ -84,43 +101,75 @@ class TopicController extends Controller
                 'old' => request()->all()]);
         }
 
-        Topic::create([
+        $topic = Topic::create([
             'name' => request('name'),
             'subject_id' => request('subject'),
         ]);
+
+        session()->flash('toast', json_encode([
+            'status' => 'Created!',
+            'message' => 'Topic: ' . $topic->name,
+            'type' => 'success'
+        ]));
 
         return response('', 200)->header('HX-Redirect', route('topics.index'));
     }
-    public function edit(Topic $topic){
-        $subjects = Subject::whereIn('course_id', $topic->subject->course()->get()->pluck('id'))->get()->pluck('name', 'id');
+    public function edit(Topic $topic)
+    {
+        $courseIds = $topic->subject->courses->pluck('id');
+        $questions_are_in_exams = $topic->questions->contains(function ($question) {
+            return $question->exams()->exists();
+        });        
 
-        $data = [
-            'topic' => $topic, 
-            'subjects' => $subjects
-        ];
-    
-        return view('topics/edit', $data);
+        $subjects = Subject::whereHas('courses', function ($query) use ($courseIds) {
+            $query->whereIn('courses.id', $courseIds);
+        })->pluck('name', 'id');
+
+        return view('topics.edit', [
+            'topic' => $topic,
+            'subjects' => $subjects,
+            'questions_are_in_exams' => $questions_are_in_exams
+        ]);
     }
 
-    public function update(Topic $topic){
+
+    public function update(Request $request, Topic $topic){
         $this->authorize('update', $topic);
 
 
-        request()->validate([
-            'name'    => ['required'],
-            'subject'     => ['required', 'integer'],
-        ]);
+        $request->validate([
+                'name' => [
+                    'required',
+                    Rule::unique('topics')->where(function ($query) use ($request) {
+                        return $query->where('subject_id', $request->subject);
+                    })->ignore($topic->id),
+                ],
+                'subject' => ['required', 'integer', 'exists:subjects,id'],
+            ]);
 
-        $topic->update([
-            'name' => request('name'),
-            'subject_id' => request('subject'),
-        ]);
+
+        $topic->update($request->all());
+    
+        session()->flash('toast', json_encode([
+            'status' => 'Updated!',
+            'message' => 'Topic: ' . $topic->name,
+            'type' => 'info'
+        ]));
 
         return redirect()->route('topics.show', $topic);
     }
     public function destroy(Topic $topic){
 
         $this->authorize('delete', $topic);
+
+        if ($topic->questions()->exists()) {
+            return back()->with('error', 'You cannot delete a topic that has questions.');
+        }
+        session()->flash('toast', json_encode([
+            'status' => 'Destroyed!',
+            'message' => 'Topic: ' . $topic->name,
+            'type' => 'warning'
+        ]));
 
         $topic->delete();
